Add meta-rule
The meta rule treats blocks of bits (uint64_t) as single cells. It
measures the bit count of a block and its neighbors. If greater than
half the bits are set in a block, then it is active.

The meta rule does not set the state of a block, it decides what
ordinary rule to apply. Currently there is one rule for active and one
for inactive.

Signed-off-by: Richard Palethorpe <richiejp@f-m.fm>